requirements.txt
openai
python-dotenv
google-generativeai
langchain-openai
langchain-google-genai
streamlit
langchain-ollama
langchain-huggingface
langsmith
langchain-aws